refactor: 위치 검색 관련 로직을 usePlacesSearch라는 hook으로 분리
